Unify StudyValidatorTest.
Issue: IBP-4735
Review: none
