+++ src/test/java/org/ibp/api/rest/dataset/validator/StudyValidatorTest.java
@@ -1,16 +1,19 @@
 package org.ibp.api.rest.dataset.validator;
 
+import org.apache.commons.lang.math.RandomUtils;
 import org.apache.commons.lang3.RandomStringUtils;
 import org.generationcp.middleware.ContextHolder;
 import org.generationcp.middleware.domain.dms.Study;
 import org.generationcp.middleware.domain.study.StudyTypeDto;
 import org.generationcp.middleware.enumeration.DatasetTypeEnum;
 import org.generationcp.middleware.manager.api.StudyDataManager;
+import org.generationcp.middleware.pojos.Person;
 import org.generationcp.middleware.pojos.workbench.Role;
 import org.generationcp.middleware.pojos.workbench.WorkbenchUser;
 import org.generationcp.middleware.service.api.study.StudyEntryService;
 import org.generationcp.middleware.service.api.study.StudyInstanceService;
 import org.generationcp.middleware.service.api.study.StudyService;
+import org.generationcp.middleware.service.api.user.UserService;
 import org.generationcp.middleware.service.impl.study.StudyInstance;
 import org.ibp.api.exception.ApiRequestValidationException;
 import org.ibp.api.exception.ForbiddenException;
@@ -34,12 +37,13 @@ import java.util.List;
 import java.util.Random;
 
 import static org.hamcrest.CoreMatchers.hasItem;
+import static org.junit.Assert.assertThat;
 import static org.mockito.Mockito.doReturn;
 
 @RunWith(MockitoJUnitRunner.class)
 public class StudyValidatorTest {
 
-	private static final int USER_ID = 10;
+	private static final Integer USER_ID = 10;
 	private static final String PROGRAM_UUID = RandomStringUtils.randomAlphabetic(10);
 	@Mock
 	private SecurityService securityService;
@@ -58,6 +62,9 @@ public class StudyValidatorTest {
 
 	@InjectMocks
 	private StudyValidator studyValidator;
+
+	@Mock
+	private UserService userService;
 
 	@Before
 	public void setup() {
@@ -225,6 +232,58 @@ public class StudyValidatorTest {
 		} catch (final ApiRequestValidationException e) {
 			Assert.assertThat(Arrays.asList(e.getErrors().get(0).getCodes()),
 				hasItem("study.has.means.dataset"));
+		}
+	}
+
+	@Test
+	public void testvalidateDeleteStudy_ThrowsException_WhenStudynotExist() {
+		try {
+			final Integer studyId = RandomUtils.nextInt();
+			this.studyValidator.validateDeleteStudy(studyId);
+		} catch (final ResourceNotFoundException e) {
+			assertThat(Arrays.asList(e.getError().getCodes()), hasItem("study.not.exist"));
+		}
+	}
+
+	@Test
+	public void testvalidateDeleteStudy_ThrowsException_WhenStudyHasProgramIUUD() {
+		try {
+			final Integer studyId = RandomUtils.nextInt();
+			final Study study = new Study();
+			study.setId(studyId);
+			study.setCreatedBy(USER_ID.toString());
+			Mockito.doReturn(study).when(studyDataManager).getStudy(studyId, false);
+			this.studyValidator.validateDeleteStudy(studyId);
+		} catch (final ApiRequestValidationException e) {
+			assertThat(Arrays.asList(e.getErrors().get(0).getCodes()), hasItem("study.template.delete.not.permitted"));
+		}
+	}
+
+	@Test
+	public void testvalidateDeleteStudy_ThrowsException_WhenTheUserIsNotOwnerOfStudy() {
+		try {
+			final Integer studyId = RandomUtils.nextInt();
+			final Study study = new Study();
+			final WorkbenchUser userLogged = new WorkbenchUser();
+			final WorkbenchUser userStudy = new WorkbenchUser();
+
+			final Person person = new Person();
+			person.setFirstName(org.apache.commons.lang.RandomStringUtils.randomAlphanumeric(8));
+			person.setMiddleName(org.apache.commons.lang.RandomStringUtils.randomAlphanumeric(1));
+			person.setLastName(org.apache.commons.lang.RandomStringUtils.randomAlphanumeric(8));
+
+			userStudy.setPerson(person);
+			userLogged.setUserid(RandomUtils.nextInt());
+			study.setId(studyId);
+			study.setProgramUUID(StudyValidatorTest.PROGRAM_UUID);
+			study.setCreatedBy(USER_ID.toString());
+			Mockito.doReturn(study).when(studyDataManager).getStudy(studyId, false);
+			Mockito.when(this.securityService.getCurrentlyLoggedInUser()).thenReturn(userLogged);
+			Mockito.when(this.userService.getUserById(study.getUser())).thenReturn(userStudy);
+
+			this.studyValidator.validateDeleteStudy(studyId);
+		} catch (final ApiRequestValidationException e) {
+			assertThat(Arrays.asList(e.getErrors().get(0).getCodes()), hasItem("study.delete.not.permitted"));
 		}
 	}
